Made final argument to CookieCreator a long

+++ src/main/kotlin/io/craigmiller160/oauth2/security/CookieCreator.kt
@@ -8,12 +8,12 @@ import java.time.format.DateTimeFormatter
 class CookieCreator {
 
     companion object {
-        private const val DEFAULT_MAX_AGE = 24 * 60 * 60
+        private const val DEFAULT_MAX_AGE = (24 * 60 * 60).toLong()
     }
 
     val cookieExpFormat: DateTimeFormatter = DateTimeFormatter.ofPattern("EEE, dd MMM yyyy HH:mm:ss z")
 
-    fun createTokenCookie(cookieName: String, cookiePath: String, token: String, maxAgeSecs: Int = DEFAULT_MAX_AGE): String {
+    fun createTokenCookie(cookieName: String, cookiePath: String, token: String, maxAgeSecs: Long = DEFAULT_MAX_AGE): String {
         val expires = ZonedDateTime.now(ZoneId.of("GMT"))
                 .plusHours(24)
         val expiresString = cookieExpFormat.format(expires)
